@@ -1,11 +1,10 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
 namespace Concert_Pal_API
@@ -14,11 +13,14 @@ namespace Concert_Pal_API
    {
       public static void Main(string[] args)
       {
-         CreateWebHostBuilder(args).Build().Run();
+         CreateHostBuilder(args).Build().Run();
       }
 
-      public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-          WebHost.CreateDefaultBuilder(args)
-              .UseStartup<Startup>();
+      public static IHostBuilder CreateHostBuilder(string[] args) =>
+          Host.CreateDefaultBuilder(args)
+              .ConfigureWebHostDefaults(webBuilder =>
+              {
+                 webBuilder.UseStartup<Startup>();
+              });
    }
 }
